Pagination User template

+++ src/AppBundle/Resources/public/js/main.js
@@ -4,17 +4,20 @@
             //var getUrl = Routing.generate('delete_user', {'id': $(this).attr('id')});
             //e.preventDefault();
             var id = $(this).attr("data-playgroup-id");
-            $.ajax({
-                type: 'delete',
-                url:  '../user/delete/'+ id,
-                data: $(this).serialize(),
-                success: function (response) {
-                    location.reload();
-                    console.log("borrado");
-                },
-                error: function (response) {
-                    console.log("no borrado");
-                }
-            });
+            if(confirm("Are you sure you want to delete this Record?")){
+                $.ajax({
+                    type: 'delete',
+                    url:  '../user/delete/'+ id,
+                    data: $(this).serialize(),
+                    success: function (response) {
+                        location.reload();
+                        console.log("borrado");
+                    },
+                    error: function (response) {
+                        console.log("no borrado");
+                    }
+                });
+            }
+            
         });
  });
